requirements.txt
Django==4.2.0 
djangorestframework==3.14.0 
django-tenants==3.5.0 
psycopg2-binary==2.9.5 
python-decouple==3.8 
django-cors-headers==4.0.0 
